work on mistakes
