+++ src/components/CharactersContextProvider/index.js
@@ -1,0 +1,3 @@
+import CharactersContextProvider from './CharactersContextProvider';
+
+export default CharactersContextProvider;
